requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
memory-profiler>=0.61.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.0.0
